AI Agent GitHub Scrapper/requirements.txt
anthropic>=0.39.0
requests>=2.31.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
